# openapi.yaml
openapi: 3.1.0
info:
  title: study-karte API
  description: >-
    AIエージェント連携型 語学学習プラットフォーム study-karte のREST API。
    学習項目（LearningItem）の登録・検索・復習期限の取得・復習結果の記録を行う。
  version: "0.1.0"
servers:
  - url: https://study-karte-oqn2yn4era-an.a.run.app
security:
  - ApiKeyAuth: []
  - FirebaseAuth: []
paths:
  /exports/legacy:
    get:
      summary: Googleログイン中の旧データ所有者が移行用JSONをダウンロードする
      responses:
        "200":
          description: 移行用JSONファイル
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyImportBundle"
        "403":
          description: Googleログインが必要
        "409":
          description: 別のユーザーが取得済み
  /imports/legacy:
    post:
      operationId: importLegacyData
      summary: Googleログイン中のユーザーへ旧データをインポートする
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyImportBundle"
      responses:
        "200":
          description: インポート結果
        "400":
          description: ファイル形式または移行トークンが不正
        "409":
          description: 別のユーザーがすでに引き取り済み
  /learning-items:
    post:
      operationId: addLearningItem
      summary: 新しく覚えたい単語・表現・文法を登録する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLearningItemInput"
      responses:
        "201":
          description: 登録済みのLearningItem
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningItem"
        "400":
          description: 入力エラー
    get:
      operationId: searchLearningItems
      summary: 登録済みのLearningItemを言語・種別で絞り込んで一覧取得する
      parameters:
        - name: language
          in: query
          schema:
            type: string
          description: "対象言語で絞り込む（例: zh-CN）"
        - name: type
          in: query
          schema:
            type: string
            enum: [vocabulary, phrase, grammar]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: LearningItemの一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearningItem"
  /due-reviews:
    get:
      operationId: getDueReviews
      summary: 現在復習すべきLearningItemを期限順に取得する
      parameters:
        - name: language
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 復習対象のLearningItem一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearningItem"
  /review-results:
    post:
      operationId: recordReviewResult
      summary: 復習結果を保存し、習熟度と次回復習日を更新する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordReviewResultInput"
      responses:
        "201":
          description: 記録されたReviewEvent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewEvent"
        "404":
          description: 対象のLearningItemが見つからない
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: ChatGPT連携の移行期間中だけ使用するサーバー間認証
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Webアプリ利用者の認証
  schemas:
    AddLearningItemInput:
      type: object
      required: [language, type, surface, meaning]
      properties:
        language:
          type: string
          description: "対象言語（例: zh-CN, en）"
        type:
          type: string
          enum: [vocabulary, phrase, grammar]
        surface:
          type: string
          description: 表記
        reading:
          type: string
          description: 発音・読み（任意）
        meaning:
          type: string
          description: ユーザー母語での意味
        note:
          type: string
          description: 補足（任意）
    LearningItem:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        language:
          type: string
        type:
          type: string
        surface:
          type: string
        reading:
          type: [string, "null"]
        meaning:
          type: string
        note:
          type: [string, "null"]
        mastery:
          type: integer
        last_reviewed_at:
          type: [string, "null"]
        next_review_at:
          type: [string, "null"]
        created_at:
          type: string
        updated_at:
          type: string
    RecordReviewResultInput:
      type: object
      required: [learning_item_id, result]
      properties:
        learning_item_id:
          type: string
        result:
          type: string
          enum: [correct, incorrect, partial, self_known]
        response_time_ms:
          type: integer
        agent:
          type: string
        quiz_type:
          type: string
    ReviewEvent:
      type: object
      properties:
        id:
          type: string
        learning_item_id:
          type: string
        user_id:
          type: string
        reviewed_at:
          type: string
        result:
          type: string
        mastery_before:
          type: integer
        mastery_after:
          type: integer
        next_review_at:
          type: string
    LegacyImportBundle:
      type: object
      required: [format, version, exported_at, claim_token, learning_items, review_events]
      properties:
        format:
          type: string
          const: study-karte-legacy-export
        version:
          type: integer
          const: 1
        exported_at:
          type: string
        claim_token:
          type: string
        learning_items:
          type: array
          items:
            $ref: "#/components/schemas/LearningItem"
        review_events:
          type: array
          items:
            $ref: "#/components/schemas/ReviewEvent"